Broke out Genome::Model::PolyphredPolyscan->setup_input into a seperate module built as a workflow operation.
Fixes in workflow to support using Workflow::Operation objects with Workflow::Simple.
Added a new flag to prevent database storage of running instances.
Fix a case where the last peer to exit at the root of a parallel workflow is parallel_index 0 and it does not rewrite input/outputs into arrays



git-svn-id: svn+ssh://svn/srv/svn/gscpan/perl_modules/trunk/Workflow@40459 127c0c0e-971a-0410-ad48-868c1a942d8d