Changes to make fastcgi view dispatcher work.
Fixed some errors in FlowCell and MutationDiagram (connecting to mgschema early).

ProcManager now exits after the request is finished, cleaning up is a waste of time when everything is read-only.  We can easily switch to rollback/clear_cache by uncommenting the lines.

Added a test that will fail if core modules dont compile, or if they cause DBI to connect (to anything) during loading.

git-svn-id: svn+ssh://svn/srv/svn/gscpan/perl_modules/trunk/Workflow@58697 127c0c0e-971a-0410-ad48-868c1a942d8d