make input and output connector optional
Currently there is a bug that when a Workflow::Model::Instance is
retrieved it never gets its input and output connector initialized. This
may be because they are supposed to be optional or it may be that they
just need to be initialized. For now I'm setting them as optional but
will need to follow up more.