Commands can now shortcut via a forked process as Events can.
If the command implements shortcut() and it returns a true value, the
operation will be deemed a success. Otherwise a job will be scheduled
and execute will be run as normal.

CHANGELOG: Commands used in workflow can now shortcut without launching
a separate LSF job first. (Notably this will make DetectVariants2 shortcutting
faster.)