CHANGELOG Enable debug of command in lsf workflow
If the environment variable WF_DEBUG_MATCH is set,
the graphical debugger ptkdb will be displayed when
that command is launched.  The interface will be
displayed as specified by the DISPLAY environment
variable, so be sure that is set correctly.
WF_DEBUG_MATCH should be set to a regex that will
match the command being bsubbed by the workflow
operation, e.g. the name of the command module.