Remove locking from workflow server.
Requires upgrade to POE::Component::IKC >= 0.2200-1

The version requirement on IKC is to allow the hub server to obtain it's own
random port instead of the parent process obtaining it and then later the
hub obtaining it from the os (which led to race conditions).

The main purpose of the locking was to block other workflow servers from
attempting to get the same ports by only letting one Server::Remote::launch
instance run at a time.  The locking code was itself vulnerable to race
conditions though.

Now, instead of locking out other workflows from launching, we just let
each workflow get it's own ports and make them communicate back to us
what port they are on.  For this communication we use a named-pipe (fifo)
that allows the hub server to communicate it's location (hostname:port)
to the parent process that can then tell the ur process.  The ur process
no longer is a server and communicates only with the hub.

The remaining locking (which appears to have been put in place to ensure that
the hub and ur processes were fully up and operational before they started to
communicate) was removed and replaced by letting the ur process announce it's
readyness to the hub server.  All messages that were intended to be sent to
the ur process (formerly a server) are sent instead through a passthru on the
hub server.  The passthru keeps the messages in an inventory until the ur
process has announced that it is ready to recieve messages.