Upgrade version of POE::Component::IKC to 0.2200-1
This is done in a non-standard way, and is a temporary solution.
In order to have code 'in the wild' that relies on the old version
of IKC we're putting it into the workflow lib directly.

Later, when all code running 'in the wild' is using this version of
IKC we'll upgrade the system installed version and remove this
installation.