Handle command inputs and calls better.
Previously, if you had a command with an OO input, you had to make
the id_by property an input, but make the object accessor NOT an input
for Workflow to dispatch the command.  This is now fixed.

Now any case where a command has properties which are id_by, or via,
etc., will work.

In cases where an output is a single value but the recipient can
take a list of values, the conversion into a one-item-list happens
automatically.

When a command is created but has errors this was not previously
checked before executing.  Now commands with errors will fail
before execution, like they do on the command-line.

Exception throwing inside of Workflow::OperationType::Command::call
did not clean up after itself, and now it does.

Error logging on the command object is now set correctly, instead of
just being set on the class twice.